+++ simple_auth/lib/src/providers/azureAD.dart
@@ -6,9 +6,8 @@ class AzureADApi extends OAuthApi {
   bool useClientSecret;
   String resource;
   AzureADApi(String identifier, String clientId, String tokenUrl, this.resource,
-      String authorizationUrl,
+      String authorizationUrl, String redirectUrl,
       {String clientSecret = "native",
-      String redirectUrl,
       List<String> scopes,
       http.Client client,
       Converter converter,
@@ -22,6 +21,7 @@ class AzureADApi extends OAuthApi {
     this.authorizationUrl = authorizationUrl;
     this.redirectUrl = redirectUrl;
     useClientSecret = clientSecret != "native";
+    this.scopes = scopes;
     this.scopesRequired = false;
   }
   @override
@@ -33,7 +33,8 @@ class AzureADApi extends OAuthApi {
       authorizationUrl,
       redirectUrl,
       resource,
-      useClientSecret);
+      useClientSecret,
+      scopes);
   @override
   Future<Map<String, String>> getRefreshTokenPostData(Account account) async {
     var map = await super.getRefreshTokenPostData(account);
@@ -55,9 +56,11 @@ class AzureADAuthenticator extends OAuthAuthenticator {
       String baseUrl,
       String redirectUrl,
       this.resource,
-      this.useClientSecret)
+      this.useClientSecret,
+      List<String> scopes)
       : super(identifier, clientId, clientSecret, tokenUrl, baseUrl,
             redirectUrl) {
+    this.scope = scopes;
     useEmbeddedBrowser = useClientSecret;
   }
 
@@ -75,6 +78,7 @@ class AzureADAuthenticator extends OAuthAuthenticator {
   Future<Map<String, dynamic>> getTokenPostData(String clientSecret) async {
     var map = await super.getTokenPostData(clientSecret);
     map["redirect_uri"] = redirectUrl;
+    map["response_type"] = "token id_token";
     if (!useClientSecret && map.containsKey("client_secret")) {
       map.remove("client_secret");
     }