+++ simple_auth_flutter_example/lib/api_definitions/azureAdTestApi.dart
@@ -3,7 +3,7 @@ import "package:http/http.dart" as http;
 
 part 'azureAdTestApi.simple_auth.dart';
 
-@AzureADApiDeclaration("AzureAdTestApi","resource","client_id",azureTennant: "azureTennant",clientSecret: "client_secret")
+@AzureADApiDeclaration("AzureAdTestApi","resource","client_id","redirecturl",azureTennant: "azureTennant",clientSecret: "client_secret")
 abstract class AzureADDefinition {
 
 }