+++ simple_auth_generator/CHANGELOG.md
@@ -1,3 +1,10 @@
+## [2.0.1] - 01-28-2019
+* Azure APIs now generate properly
+
+## [2.0] - 11-09-2018
+* The rest of the providers now require a redirect
+* Cleaned up dependencies
+
 ## [1.0.3] - 08-30-2018
 * Updated dependencies
 ## [1.0.2] - 08-30-2018